#286 Updated Laratrust - was missing some upgrade notes.

+++ database/seeds/LaratrustSeeder.php
@@ -2,6 +2,8 @@
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Config;
 
 class LaratrustSeeder extends Seeder
 {
@@ -15,14 +17,13 @@ class LaratrustSeeder extends Seeder
         $this->command->info('Truncating User, Role and Permission tables');
         $this->truncateLaratrustTables();
 
-        $config = config('laratrust_seeder.role_structure');
-        $userPermission = config('laratrust_seeder.permission_structure');
+        $config = config('laratrust_seeder.roles_structure');
         $mapPermission = collect(config('laratrust_seeder.permissions_map'));
 
         foreach ($config as $key => $modules) {
 
             // Create a new role
-            $role = \App\Role::create([
+            $role = \App\Role::firstOrCreate([
                 'name' => $key,
                 'display_name' => ucwords(str_replace('_', ' ', $key)),
                 'description' => ucwords(str_replace('_', ' ', $key))
@@ -51,51 +52,17 @@ class LaratrustSeeder extends Seeder
             // Attach all permissions to the role
             $role->permissions()->sync($permissions);
 
-            $this->command->info("Creating '{$key}' user");
-
-            // Create default user for each role
-            $user = \App\User::create([
-                'name' => ucwords(str_replace('_', ' ', $key)),
-                'email' => $key.'@app.com',
-                'password' => bcrypt('password')
-            ]);
-
-            $user->attachRole($role);
-        }
-
-        // Creating user with permissions
-        if (!empty($userPermission)) {
-
-            foreach ($userPermission as $key => $modules) {
-
-                foreach ($modules as $module => $value) {
-
-                    // Create default user for each permission set
-                    $user = \App\User::create([
-                        'name' => ucwords(str_replace('_', ' ', $key)),
-                        'email' => $key.'@app.com',
-                        'password' => bcrypt('password'),
-                        'remember_token' => str_random(10),
-                    ]);
-                    $permissions = [];
-
-                    foreach (explode(',', $value) as $p => $perm) {
-
-                        $permissionValue = $mapPermission->get($perm);
-
-                        $permissions[] = \App\Permission::firstOrCreate([
-                            'name' => $permissionValue . '-' . $module,
-                            'display_name' => ucfirst($permissionValue) . ' ' . ucfirst($module),
-                            'description' => ucfirst($permissionValue) . ' ' . ucfirst($module),
-                        ])->id;
-
-                        $this->command->info('Creating Permission to '.$permissionValue.' for '. $module);
-                    }
-                }
-
-                // Attach all permissions to the user
-                $user->permissions()->sync($permissions);
+            if(Config::get('laratrust_seeder.create_users')) {
+                $this->command->info("Creating '{$key}' user");
+                // Create default user for each role
+                $user = \App\User::create([
+                    'name' => ucwords(str_replace('_', ' ', $key)),
+                    'email' => $key.'@app.com',
+                    'password' => bcrypt('password')
+                ]);
+                $user->attachRole($role);
             }
+
         }
     }
 
@@ -110,9 +77,13 @@ class LaratrustSeeder extends Seeder
         DB::table('permission_role')->truncate();
         DB::table('permission_user')->truncate();
         DB::table('role_user')->truncate();
-        \App\User::truncate();
-        \App\Role::truncate();
-        \App\Permission::truncate();
+        if(Config::get('laratrust_seeder.truncate_tables')) {
+            \App\Role::truncate();
+            \App\Permission::truncate();
+        }
+        if(Config::get('laratrust_seeder.truncate_tables') && Config::get('laratrust_seeder.create_users')) {
+            \App\User::truncate();
+        }
         Schema::enableForeignKeyConstraints();
     }
 }
